# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Storyblok Voice Assistant API
  description: |
    Voice-enabled content discovery for Storyblok using AWS Bedrock and natural conversation.
    
    ## Features
    - Natural language conversation with Claude AI
    - Semantic content search via Storyblok Strata
    - Dynamic result limits based on user intent (e.g., "first 5 articles")
    - Multi-turn conversation support
    - Accessible voice and text interfaces
    
    ## Base URL
    - Development: `http://localhost:8000`
    
  version: 1.0.0
  contact:
    name: Storyblok Voice Assistant
  license:
    name: Hackathon Project
servers:
  - url: http://localhost:8000
    description: Development server
...
paths:
...
  /api/test-bedrock:
    get:
      tags:
        - Debug
      summary: Test Bedrock connection
      description: |
        Test endpoint for AWS Bedrock integration.
        Only available when DEBUG=true in environment configuration.
      operationId: testBedrock
      responses:
        '200':
          description: Test completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success, error]
                  response:
                    type: object
                    description: Bedrock response (on success)
                  error:
                    type: string
                    description: Error message (on error)
        '404':
          description: Debug mode disabled
  /api/test-storyblok:
    get:
      tags:
        - Debug
      summary: Test Storyblok connection
      description: |
        Test endpoint for Storyblok Strata integration.
        Only available when DEBUG=true in environment configuration.
      operationId: testStoryblok
      parameters:
        - name: term
          in: query
          description: Search term to test
          required: false
          schema:
            type: string
            default: test
      responses:
        '200':
          description: Test completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success, error]
                  results:
                    $ref: '#/components/schemas/SearchResults'
                  error:
                    type: string
        '404':
          description: Debug mode disabled
  /api/story/{story_id}:
    get:
      tags:
        - Stories
      summary: Get full story by ID
      description: |
        Fetch complete story details from Storyblok API by story ID.
        Returns full story data including content, metadata, and all fields.
      operationId: getFullStory
      parameters:
        - name: story_id
          in: path
          description: Storyblok story ID
          required: true
          schema:
            type: integer
            example: 12345
      responses:
        '200':
          description: Story found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  story:
                    type: object
                    description: Complete story data from Storyblok
        '404':
          description: Story not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Story with ID 12345 not found"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
components:
  schemas:
...
    StoryResult:
      type: object
      description: Story result from Storyblok Strata API
      required:
        - body
        - cursor
        - name
        - slug
        - story_id
      properties:
       body:
        type: string
        description: Story body/content as text from Strata API
        example: "This is a comprehensive guide to modern web development..."
       cursor:
        type: integer
        description: Cursor position for pagination
        example: 0
       name:
        type: string
        description: Story name/title
        example: "Web Development Guide 2024"
       slug:
        type: string
        description: Story slug/path
        example: "web-development-guide-2024"
       story_id:
        type: integer
        description: Unique Storyblok story identifier
        example: 123456
       content_type:
        type: string
        nullable: true
        description: Content type/component of the story (e.g., article, blog_post, page)
        example: "article"
       full_story:
          type: object
          nullable: true
          description: |
            Full story data fetched from Storyblok API (when available).
            Contains complete story information including content, metadata, and all fields.
          properties:
           id:
            type: integer
            description: Story ID
           name:
            type: string
            description: Story name
           content:
            type: object
            description: Story content structure
           content_type:
            type: string
            description: Story content type
           published_at:
            type: string
            format: date-time
            description: Publication timestamp
           created_at:
            type: string
            format: date-time
            description: Creation timestamp
    SearchResults:
      type: object
      properties:
        stories:
          type: array
          items:
            $ref: '#/components/schemas/StoryResult'
          description: Array of story results
        total:
          type: integer
          description: Total number of results
          example: 15
...